Kill UltravoxSessionState, as in other SDKs
