+++ .idea/workspace.xml
@@ -12,26 +12,16 @@
       </layout>
     </layouts>
   </component>
-  <component name="AndroidLayouts">
-    <shared>
-      <config />
-    </shared>
-    <layouts>
-      <layout url="file://$PROJECT_DIR$/demoapp/src/main/res/layout/activity_main.xml">
-        <config>
-          <theme>@style/Theme.UltravoxClient</theme>
-        </config>
-      </layout>
-    </layouts>
-  </component>
   <component name="AutoImportSettings">
     <option name="autoReloadType" value="NONE" />
   </component>
   <component name="ChangeListManager">
     <list default="true" id="0e7e9986-8331-4ae5-8c43-d2b86550bb12" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/.idea/runConfigurations/publishToMaven.xml" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/local.properties" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ultravox_client/build.gradle.kts" beforeDir="false" afterPath="$PROJECT_DIR$/ultravox_client/build.gradle.kts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/demoapp/src/main/java/ai/ultravox/demoapp/MainActivity.kt" beforeDir="false" afterPath="$PROJECT_DIR$/demoapp/src/main/java/ai/ultravox/demoapp/MainActivity.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ultravox_client/src/main/java/ai/ultravox/UltravoxSession.kt" beforeDir="false" afterPath="$PROJECT_DIR$/ultravox_client/src/main/java/ai/ultravox/UltravoxSession.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ultravox_client/src/main/java/ai/ultravox/UltravoxSessionState.kt" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -65,12 +55,6 @@
   <component name="ProjectColorInfo">{
   &quot;associatedIndex&quot;: 5
 }</component>
-  <component name="ProblemsViewState">
-    <option name="selectedTabId" value="CurrentFile" />
-  </component>
-  <component name="ProjectColorInfo">{
-  &quot;associatedIndex&quot;: 5
-}</component>
   <component name="ProjectId" id="2lCoVzUzrotJrtAZH1VeVHk7mx9" />
   <component name="ProjectLevelVcsManager">
     <ConfirmationsSetting value="2" id="Add" />
@@ -90,8 +74,7 @@
     "cf.first.check.clang-format": "false",
     "cidr.known.project.marker": "true",
     "dart.analysis.tool.window.visible": "false",
-    "git-widget-placeholder": "mike/demo-app",
-    "git-widget-placeholder": "mike/demo-app",
+    "git-widget-placeholder": "main",
     "kotlin-language-version-configured": "true",
     "project.structure.last.edited": "Suggestions",
     "project.structure.proportion": "0.17",
