+++ ultravox_client/src/main/java/ai/ultravox/UltravoxSession.kt
@@ -1,10 +1,13 @@
 package ai.ultravox
 
 import android.content.Context
+import android.os.Build
+import android.util.Log
 import io.livekit.android.LiveKit
 import io.livekit.android.events.RoomEvent
 import io.livekit.android.events.collect
 import io.livekit.android.room.Room
+import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import okhttp3.HttpUrl.Companion.toHttpUrl
@@ -14,6 +17,8 @@ import okhttp3.WebSocket
 import okhttp3.WebSocketListener
 import org.json.JSONObject
 
+typealias UltravoxSessionListener = () -> Unit
+
 @Suppress("unused")
 class UltravoxSession(
     ctx: Context,
@@ -21,15 +26,51 @@ class UltravoxSession(
     private val client: OkHttpClient = OkHttpClient(),
     private val experimentalMessages: Set<String> = HashSet(),
 ) {
-    private val state = UltravoxSessionState()
     private var socket: WebSocket? = null
     private var room: Room = LiveKit.create(ctx)
 
-    fun joinCall(joinUrl: String): UltravoxSessionState {
-        if (state.status != UltravoxSessionStatus.DISCONNECTED) {
+    private val _transcripts = ArrayList<Transcript>()
+
+    val transcripts
+        get() = _transcripts.toImmutableList()
+
+    val lastTranscript
+        get() = _transcripts.lastOrNull()
+
+    var status = UltravoxSessionStatus.DISCONNECTED
+        private set(value) {
+            val prev = field
+            field = value
+            if (prev != value) {
+                fireListeners("status")
+            }
+        }
+
+    @Suppress("MemberVisibilityCanBePrivate")
+    var lastExperimentalMessage: JSONObject? = null
+        private set(value) {
+            field = value
+            fireListeners("experimental_message")
+        }
+
+    private val listeners = HashMap<String, ArrayList<UltravoxSessionListener>>()
+
+    fun listen(event: String, listener: UltravoxSessionListener) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            listeners.putIfAbsent(event, ArrayList())
+        } else {
+            if (!listeners.containsKey("event")) {
+                listeners[event] = ArrayList()
+            }
+        }
+        listeners[event]!!.add(listener)
+    }
+
+    fun joinCall(joinUrl: String) {
+        if (status != UltravoxSessionStatus.DISCONNECTED) {
             throw RuntimeException("Cannot join a new call while already in a call")
         }
-        state.status = UltravoxSessionStatus.CONNECTING
+        status = UltravoxSessionStatus.CONNECTING
         var httpUrl = if (joinUrl.startsWith("wss://") or joinUrl.startsWith("ws://")) {
             // This is the expected case, but OkHttp expects http(s) protocol even
             // for WebSocket requests for some reason.
@@ -58,7 +99,7 @@ class UltravoxSession(
                         }
                         room.connect(message["roomUrl"] as String, message["token"] as String)
                         room.localParticipant.setMicrophoneEnabled(true)
-                        state.status = UltravoxSessionStatus.IDLE
+                        status = UltravoxSessionStatus.IDLE
                     }
                 }
             }
@@ -71,8 +112,6 @@ class UltravoxSession(
                 disconnect()
             }
         })
-
-        return state
     }
 
     fun leaveCall() {
@@ -80,8 +119,8 @@ class UltravoxSession(
     }
 
     fun sendText(text: String) {
-        if (!state.status.live) {
-            throw RuntimeException("Cannot send text while not connected. Current status is " + state.status + ".")
+        if (!status.live) {
+            throw RuntimeException("Cannot send text while not connected. Current status is $status.")
         }
         val message = JSONObject()
         message.put("type", "input_text_message")
@@ -90,13 +129,13 @@ class UltravoxSession(
     }
 
     private fun disconnect() {
-        if (state.status == UltravoxSessionStatus.DISCONNECTED) {
+        if (status == UltravoxSessionStatus.DISCONNECTED) {
             return
         }
-        state.status = UltravoxSessionStatus.DISCONNECTING
+        status = UltravoxSessionStatus.DISCONNECTING
         room.disconnect()
         socket?.cancel()
-        state.status = UltravoxSessionStatus.DISCONNECTED
+        status = UltravoxSessionStatus.DISCONNECTED
     }
 
     private fun onDataReceived(event: RoomEvent.DataReceived) {
@@ -104,9 +143,9 @@ class UltravoxSession(
         when (message["type"]) {
             "state" -> {
                 when (message["state"]) {
-                    "listening" -> state.status = UltravoxSessionStatus.LISTENING
-                    "thinking" -> state.status = UltravoxSessionStatus.THINKING
-                    "speaking" -> state.status = UltravoxSessionStatus.SPEAKING
+                    "listening" -> status = UltravoxSessionStatus.LISTENING
+                    "thinking" -> status = UltravoxSessionStatus.THINKING
+                    "speaking" -> status = UltravoxSessionStatus.SPEAKING
                 }
             }
 
@@ -114,7 +153,7 @@ class UltravoxSession(
                 val transcript = message["transcript"] as JSONObject
                 val medium =
                     if (transcript.has("medium") && transcript["medium"] == "text") Transcript.Medium.TEXT else Transcript.Medium.VOICE
-                state.addOrUpdateTranscript(
+                addOrUpdateTranscript(
                     Transcript(
                         transcript["text"] as String,
                         transcript["final"] as Boolean,
@@ -128,7 +167,7 @@ class UltravoxSession(
                 val medium =
                     if (message["type"] == "agent_text_transcript") Transcript.Medium.TEXT else Transcript.Medium.VOICE
                 if (message.has("text") && message["text"] != JSONObject.NULL) {
-                    state.addOrUpdateTranscript(
+                    addOrUpdateTranscript(
                         Transcript(
                             message["text"] as String,
                             message["final"] as Boolean,
@@ -137,9 +176,9 @@ class UltravoxSession(
                         )
                     )
                 } else if (message.has("delta") && message["delta"] != JSONObject.NULL) {
-                    val last = state.lastTranscript
+                    val last = lastTranscript
                     if (last != null && last.speaker == Transcript.Role.AGENT) {
-                        state.addOrUpdateTranscript(
+                        addOrUpdateTranscript(
                             Transcript(
                                 last.text + message["delta"] as String,
                                 message["final"] as Boolean,
@@ -153,10 +192,19 @@ class UltravoxSession(
 
             else -> {
                 if (experimentalMessages.isNotEmpty()) {
-                    state.lastExperimentalMessage = message
+                    lastExperimentalMessage = message
                 }
             }
         }
+    }
+
+    private fun addOrUpdateTranscript(transcript: Transcript) {
+        val last = lastTranscript
+        if (last != null && !last.isFinal && last.speaker == transcript.speaker) {
+            _transcripts.removeLast()
+        }
+        _transcripts.add(transcript)
+        fireListeners("transcript")
     }
 
     private fun sendData(message: JSONObject) {
@@ -165,4 +213,16 @@ class UltravoxSession(
         }
     }
 
+    private fun fireListeners(event: String) {
+        if (!listeners.containsKey(event)) {
+            return
+        }
+        for (listener in listeners[event]!!) {
+            try {
+                listener()
+            } catch (e: Exception) {
+                Log.w("UltravoxClient", "Listener error: ", e)
+            }
+        }
+    }
 }
